Make changes based on suggestions, created parser.[ch], io_header.[ch], utils.[ch], mib_header.h
