+++ include/mib_header.h
@@ -1,4 +1,40 @@
 // clang-format Language: C
+#include "mib_macros.h"
+#ifndef MIB_HEADER_DAC_RX_H
+#define MIB_HEADER_DAC_RX_H
+
+typedef struct {
+  char dac_format[4];
+  unsigned int threshold0;
+  unsigned int threshold1;
+  unsigned int threshold2;
+  unsigned int threshold3;
+  unsigned int threshold4;
+  unsigned int threshold5;
+  unsigned int threshold6;
+  unsigned int threshold7;
+  unsigned int preamp;
+  unsigned int ikrum;
+  unsigned int shaper;
+  unsigned int disc;
+  unsigned int disc_LS;
+  unsigned int shaper_test;
+  unsigned int dac_disc_L;
+  unsigned int dac_test;
+  unsigned int dac_disc_H;
+  unsigned int delay;
+  unsigned int TP_buff_in;
+  unsigned int TP_buff_out;
+  unsigned int RPZ;
+  unsigned int GND;
+  unsigned int TP_ref;
+  unsigned int FBK;
+  unsigned int Cas;
+  unsigned int TP_ref_A;
+  unsigned int TP_ref_B;
+} dac_rx;
+
+#endif
 /*
 field 1: header ID (MQ1)
 field 2: sequence number of a frame
@@ -27,8 +63,6 @@ field  54     / 138  : the pixel data bit depth
 
 The remainings are null bytes for padding.
 */
-#include "mib_header_DAC.h"
-#include "mib_macros.h"
 
 #ifndef MIB_HEADER_MQ1_SINGLE_H
 #define MIB_HEADER_MQ1_SINGLE_H
