--- include/mq1_single.h
@@ -1,12 +1,0 @@
-// clang-format Language: C
-#include "mib_header_MQ1.h"
-
-#ifndef PARSE_MQ1_SINGLE_H
-#define PARSE_MQ1_SINGLE_H
-void parse_mq1_single(const char *header, mq1s *mq1_single);
-#endif
-
-#ifndef PRINT_SINGLE_MIB_HEADER_H
-#define PRINT_SINGLE_MIB_HEADER_H
-void print_single_mib_header(mq1s header);
-#endif
