+++ src/utils.c
@@ -1,4 +1,4 @@
-#include "mib_utils.h"
+#include "utils.h"
 
 #include <stdio.h>
 #include <stdlib.h>
